Extract semantic classes to separate calculator
