+++ src/lm_polygraph/estimators/semantic_entropy.py
@@ -22,8 +22,7 @@ class SemanticEntropy(Estimator):
             [
                 "sample_log_probs",
                 "sample_texts",
-                "semantic_matrix_entail",
-                "entailment_id",
+                "semantic_classes_entail",
             ],
             "sequence",
         )
@@ -47,8 +46,8 @@ class SemanticEntropy(Estimator):
         """
         loglikelihoods_list = stats["sample_log_probs"]
 
-        # entailment_id = stats["deberta"].deberta.config.label2id["ENTAILMENT"] # TODO: Why this is here??
-        self._is_entailment = stats["semantic_matrix_classes"] == stats["entailment_id"]
+        self._class_to_sample = stats["semantic_classes_entail"]["class_to_sample"]
+        self._sample_to_class = stats["semantic_classes_entail"]["sample_to_class"]
 
         # Concatenate hypos with input texts
         hyps_list = [[] for _ in stats["input_texts"]]
@@ -66,8 +65,6 @@ class SemanticEntropy(Estimator):
     ) -> np.array:
         if log_weights is None:
             log_weights = [None for _ in hyps_list]
-
-        self.get_classes(hyps_list)
 
         semantic_logits = {}
         # Iteration over batch
@@ -88,41 +85,3 @@ class SemanticEntropy(Estimator):
                 ]
             )
         return np.array([semantic_logits[i] for i in range(len(hyps_list))])
-
-    def get_classes(self, hyps_list: List[List[str]]):
-        self._sample_to_class = {}
-        self._class_to_sample: Dict[int, List] = defaultdict(list)
-
-        [
-            self._determine_class(idx, i)
-            for idx, hyp in enumerate(hyps_list)
-            for i in range(len(hyp))
-        ]
-
-        return self._sample_to_class, self._class_to_sample
-
-    def _determine_class(self, idx: int, i: int):
-        # For first hypo just create a zeroth class
-        if i == 0:
-            self._class_to_sample[idx] = [[0]]
-            self._sample_to_class[idx] = {0: 0}
-
-            return 0
-
-        # Iterate over existing classes and return if hypo belongs to one of them
-        for class_id in range(len(self._class_to_sample[idx])):
-            class_text_id = self._class_to_sample[idx][class_id][0]
-            forward_entailment = self._is_entailment[idx, class_text_id, i]
-            backward_entailment = self._is_entailment[idx, i, class_text_id]
-            if forward_entailment and backward_entailment:
-                self._class_to_sample[idx][class_id].append(i)
-                self._sample_to_class[idx][i] = class_id
-
-                return class_id
-
-        # If none of the existing classes satisfy - create new one
-        new_class_id = len(self._class_to_sample[idx])
-        self._sample_to_class[idx][i] = new_class_id
-        self._class_to_sample[idx].append([i])
-
-        return new_class_id
